添加-e参数

+++ main.go
@@ -24,25 +24,41 @@ func main() {
 		}
 	}
 
-	// Always open a new CMD window for the interactive selection.
-	// Write commands to a temp .bat file to avoid inline quoting issues with cmd.exe.
+	// Check for -e/--explorer flag (support any position)
+	hasExplorer := false
+	var filteredArgs []string
+	for _, a := range os.Args[1:] {
+		if a == "-e" || a == "--explorer" {
+			hasExplorer = true
+		} else {
+			filteredArgs = append(filteredArgs, a)
+		}
+	}
+
 	exe, err := os.Executable()
 	if err != nil {
 		fmt.Fprintln(os.Stderr, "Error:", err)
 		os.Exit(1)
 	}
 
-	tmpPath := filepath.Join(os.TempDir(), "jumpd_path.txt")
 	tmpBat := filepath.Join(os.TempDir(), "jumpd_run.bat")
-	os.Remove(tmpPath)
 	os.Remove(tmpBat)
 
-	args := os.Args[1:]
-	// Use %%%%i in Sprintf to produce %%i in the bat file (Go %% -> %, so %%%% -> %%)
-	bat := fmt.Sprintf(
-		"@echo off\r\n\"%s\" --wrapper %s\r\nif exist \"%s\" (for /f \"usebackq\" %%%%i in (\"%s\") do (cd /d \"%%%%i\" & for %%%%j in (\"%%%%~nxi\") do title %%%%j & cls)) else exit\r\n",
-		exe, strings.Join(args, " "), tmpPath, tmpPath,
-	)
+	var bat string
+	if hasExplorer {
+		// Explorer mode: after selection, CMD window exits automatically
+		bat = fmt.Sprintf("@echo off\r\n\"%s\" --wrapper --explorer %s\r\nexit\r\n",
+			exe, strings.Join(filteredArgs, " "))
+	} else {
+		// Default: cd into selected directory
+		tmpPath := filepath.Join(os.TempDir(), "jumpd_path.txt")
+		os.Remove(tmpPath)
+		bat = fmt.Sprintf(
+			"@echo off\r\n\"%s\" --wrapper %s\r\nif exist \"%s\" (for /f \"usebackq\" %%%%i in (\"%s\") do (cd /d \"%%%%i\" & for %%%%j in (\"%%%%~nxi\") do title %%%%j & cls)) else exit\r\n",
+			exe, strings.Join(filteredArgs, " "), tmpPath, tmpPath,
+		)
+	}
+
 	if err := os.WriteFile(tmpBat, []byte(bat), 0644); err != nil {
 		fmt.Fprintln(os.Stderr, "Error:", err)
 		os.Exit(1)
